Flytter felles funksjon til egen fil. Uryddig at OMS benytter util fra BP.
EY-2826
