+++ apps/etterlatte-vilkaarsvurdering/src/main/kotlin/vilkaarsvurdering/vilkaar/BarnepensjonVilkaar1967.kt
@@ -3,17 +3,14 @@ package no.nav.etterlatte.vilkaarsvurdering.vilkaar
 import no.nav.etterlatte.funksjonsbrytere.FeatureToggleService
 import no.nav.etterlatte.libs.common.behandling.Virkningstidspunkt
 import no.nav.etterlatte.libs.common.grunnlag.Grunnlag
-import no.nav.etterlatte.libs.common.grunnlag.Opplysning
 import no.nav.etterlatte.libs.common.grunnlag.hentDoedsdato
 import no.nav.etterlatte.libs.common.grunnlag.hentFoedselsdato
 import no.nav.etterlatte.libs.common.vilkaarsvurdering.Delvilkaar
 import no.nav.etterlatte.libs.common.vilkaarsvurdering.Lovreferanse
 import no.nav.etterlatte.libs.common.vilkaarsvurdering.Vilkaar
-import no.nav.etterlatte.libs.common.vilkaarsvurdering.VilkaarOpplysningType
 import no.nav.etterlatte.libs.common.vilkaarsvurdering.VilkaarOpplysningType.AVDOED_DOEDSDATO
 import no.nav.etterlatte.libs.common.vilkaarsvurdering.VilkaarOpplysningType.SOEKER_FOEDSELSDATO
 import no.nav.etterlatte.libs.common.vilkaarsvurdering.VilkaarType
-import no.nav.etterlatte.libs.common.vilkaarsvurdering.Vilkaarsgrunnlag
 import no.nav.etterlatte.vilkaarsvurdering.VilkaarFeatureToggle
 
 object BarnepensjonVilkaar1967 {
@@ -406,13 +403,5 @@ object BarnepensjonVilkaar1967 {
                     ledd = 1,
                     bokstav = "a",
                 ),
-        )
-
-    internal fun <T> Opplysning.Konstant<out T?>.toVilkaarsgrunnlag(type: VilkaarOpplysningType) =
-        Vilkaarsgrunnlag(
-            id = id,
-            opplysningsType = type,
-            kilde = kilde,
-            opplysning = verdi,
         )
 }
